change type database

+++ database/migrations/2024_08_22_021046_create_riwayats_table.php
@@ -12,7 +12,7 @@ return new class extends Migration {
     {
         Schema::create('riwayats', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('no_assets');
+            $table->unsignedBigInteger('no_assets');
             $table->unsignedInteger('idUser');
             $table->string('StatusAwal');
             $table->string('StatusAkhir');
